packages moves + sample conf for HBAse
