--- logisland-plugins/logisland-hbase-plugin/src/test/java/com/hurence/logisland/hbase/util/TestObjectSerDe.java
@@ -1,56 +1,0 @@
-
-package com.hurence.logisland.hbase.util;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-
-public class TestObjectSerDe {
-
-   @Test
-    public void testDeserializeSuccessful() throws IOException {
-       final ObjectSerDe serDe = new ObjectSerDe();
-
-       final String myObject = "myObject";
-       final ByteArrayOutputStream bOut = new ByteArrayOutputStream();
-       final ObjectOutputStream out = new ObjectOutputStream(bOut);
-       out.writeObject(myObject);
-
-       byte[] myObjectBytes = bOut.toByteArray();
-       Assert.assertNotNull(myObjectBytes);
-       Assert.assertTrue(myObjectBytes.length > 0);
-
-       final Object deserialized = serDe.deserialize(myObjectBytes);
-       Assert.assertTrue(deserialized instanceof String);
-       Assert.assertEquals(myObject, deserialized);
-   }
-
-    @Test
-    public void testDeserializeNull() throws IOException {
-        final ObjectSerDe serDe = new ObjectSerDe();
-        final Object deserialized = serDe.deserialize(null);
-        Assert.assertNull(deserialized);
-    }
-
-    @Test
-    public void testSerialize() throws IOException, ClassNotFoundException {
-        final ByteArrayOutputStream out = new ByteArrayOutputStream();
-        final String myObject = "myObject";
-
-        final ObjectSerDe serDe = new ObjectSerDe();
-        serDe.serialize(myObject, out);
-
-        final ByteArrayInputStream bIn = new ByteArrayInputStream(out.toByteArray());
-        final ObjectInputStream in = new ObjectInputStream(bIn);
-
-        final Object deserialized = in.readObject();
-        Assert.assertTrue(deserialized instanceof String);
-        Assert.assertEquals(myObject, deserialized);
-    }
-
-}
