--- logisland-services/logisland-hbase-client-service-api/src/main/java/com/hurence/logisland/hbase/scan/ResultHandler.java
@@ -1,10 +1,0 @@
-package com.hurence.logisland.hbase.scan;
-
-/**
- * Handles a single row from an HBase scan.
- */
-public interface ResultHandler {
-
-    void handle(byte[] row, ResultCell[] resultCells);
-
-}
