+++ logisland-services/logisland-hbase-client-service-api/src/main/java/com/hurence/logisland/processor/hbase/scan/ResultHandler.java
@@ -1,0 +1,10 @@
+package com.hurence.logisland.processor.hbase.scan;
+
+/**
+ * Handles a single row from an HBase scan.
+ */
+public interface ResultHandler {
+
+    void handle(byte[] row, ResultCell[] resultCells);
+
+}
