--- logisland-services/logisland-hbase_1_1_2-client-service/src/test/java/com/hurence/logisland/hbase/TestProcessor.java
@@ -1,40 +1,0 @@
-
-package com.hurence.logisland.hbase;
-
-import com.hurence.logisland.component.PropertyDescriptor;
-import com.hurence.logisland.processor.AbstractProcessor;
-import com.hurence.logisland.processor.ProcessContext;
-import com.hurence.logisland.processor.ProcessException;
-import com.hurence.logisland.record.Record;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-
-
-
-public class TestProcessor extends AbstractProcessor {
-
-    static final PropertyDescriptor HBASE_CLIENT_SERVICE = new PropertyDescriptor.Builder()
-            .name("hbase.client.service")
-            .description("HBaseClientService")
-            .identifiesControllerService(HBaseClientService.class)
-            .required(true)
-            .build();
-
-
-    @Override
-    public List<PropertyDescriptor> getSupportedPropertyDescriptors() {
-        List<PropertyDescriptor> propDescs = new ArrayList<>();
-        propDescs.add(HBASE_CLIENT_SERVICE);
-        return propDescs;
-    }
-
-    @Override
-    public Collection<Record> process(ProcessContext context, Collection<Record> records) {
-        return Collections.emptyList();
-    }
-
-
-}
